+++ HISTORY.rst
@@ -15,6 +15,7 @@ History
     * Only try to apply security filter query if model is subclass of LocalRolesMixin (rudaporto).
     * Change BaseResource.default_filters hook to be a method: it receives and returns a query (with filters or params applied) (rudaporto).
     * Pin pyramid version to 1.7.3 (rudaporto).
+    * Change create_filter_from_query_params to not use utils.data.native_value to convert string to integer (rudaporto).
 
 
 1.1.1 (2016-10-04)